refactor: `ImageListResponse.type.ts` -> `image.type.ts`로 이름 변경
