+++ src/drawer/types/image.type.ts
